+++ packages/web-core/src/services/conversation.service.ts
@@ -35,6 +35,8 @@ export class ConversationService<
   private conversationId: string;
   private participants: string[];
   private addCharacters?: (names: string[]) => Promise<void>;
+  private startRecording?: () => Promise<void>;
+  private stopRecording?: () => void;
   private packetQueue: PacketQueueItem<InworldPacketT>[] = [];
   private ttsPlaybackAction = TtsPlaybackAction.UNKNOWN;
 
@@ -44,10 +46,14 @@ export class ConversationService<
       participants,
       conversationId,
       addCharacters,
+      startRecording,
+      stopRecording,
     }: {
       participants: string[];
       conversationId?: string;
       addCharacters: (names: string[]) => Promise<void>;
+      startRecording?: () => Promise<void>;
+      stopRecording?: () => void;
     },
   ) {
     this.connection = connection;
@@ -55,6 +61,8 @@ export class ConversationService<
     this.participants = participants;
 
     this.addCharacters = addCharacters;
+    this.startRecording = startRecording;
+    this.stopRecording = stopRecording;
   }
 
   getConversationId() {
@@ -110,7 +118,11 @@ export class ConversationService<
     // If audio session is started, we need to end it before updating participants
     if (this.connection.getAudioSessionAction() === AudioSessionState.START) {
       needToReacreateAudioSession = true;
-      await this.sendAudioSessionEnd();
+      this.stopRecording?.();
+      this.beforeAudioSessionEnd();
+      await this.connection.send(() =>
+        this.connection.getEventFactory().audioSessionEnd({ conversationId }),
+      );
     }
 
     // Load characters if they are not loaded
@@ -143,7 +155,11 @@ export class ConversationService<
 
     // If audio session was started before, we need to restart it
     if (needToReacreateAudioSession) {
-      await this.sendAudioSessionStart();
+      this.beforeAudioSessionStart();
+      await this.connection.send(() =>
+        this.connection.getEventFactory().audioSessionStart({ conversationId }),
+      );
+      this.startRecording?.();
     }
 
     await this.resolveInterval(
@@ -204,8 +220,7 @@ export class ConversationService<
       throw Error('Audio session is already started');
     }
 
-    this.connection.setAudioSessionAction(AudioSessionState.START);
-    this.connection.setCurrentAudioConversation(this);
+    this.beforeAudioSessionStart();
 
     return this.ensureConversation(() =>
       this.connection.getEventFactory().audioSessionStart({
@@ -226,8 +241,7 @@ export class ConversationService<
     }
 
     return this.ensureConversation(() => {
-      this.connection.setAudioSessionAction(AudioSessionState.END);
-      this.connection.setCurrentAudioConversation(undefined);
+      this.beforeAudioSessionEnd();
 
       return this.connection
         .getEventFactory()
@@ -351,6 +365,16 @@ export class ConversationService<
     this.releaseQueue();
 
     return sent;
+  }
+
+  private beforeAudioSessionStart() {
+    this.connection.setAudioSessionAction(AudioSessionState.START);
+    this.connection.setCurrentAudioConversation(this);
+  }
+
+  private beforeAudioSessionEnd() {
+    this.connection.setAudioSessionAction(AudioSessionState.END);
+    this.connection.setCurrentAudioConversation(undefined);
   }
 
   setTtsPlaybackAction(action: TtsPlaybackAction) {